restructure youtube service
